♻️ Refactor CheckboxGroup to new folder
