+++ src/components/checkbox/Checkbox.mdx
@@ -14,10 +14,6 @@ import { Checkbox } from '.'
 
 <Canvas of={CheckboxStories.SizesOfCheckbox} />
 
-### 2. Checkbox Group
-
-<Canvas of={CheckboxStories.GroupOfCheckbox} />
-
 ## API
 
 <ArgTypes of={CheckboxStories} />
