+++ src/components/checkbox/Checkbox.types.ts
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react'
 
-type CheckboxSize = 'small' | 'medium' | 'large'
+export type CheckboxSize = 'small' | 'medium' | 'large'
 
 export interface CheckboxProps {
   id: string
@@ -11,21 +11,4 @@ export interface CheckboxProps {
   checked?: boolean
   onChange?: (checked: boolean) => void
   disabled?: boolean
-}
-
-export type OptionType = {
-  id: string
-  value: string
-  label: string
-  disabled?: boolean
-}
-
-export interface CheckboxGroupProps {
-  options: OptionType[]
-  name: string
-  label?: string
-  size?: CheckboxSize
-  direction?: 'horizontal' | 'vertical'
-  checkedOptions?: string[]
-  onChange?: (checkedOptions: string[]) => void
 }
