+++ src/components/checkbox/index.tsx
@@ -1,3 +1,2 @@
 export { Checkbox } from './Checkbox'
-export { CheckboxGroup } from './CheckboxGroup'
-export type { CheckboxProps, CheckboxGroupProps } from './Checkbox.types'
+export type { CheckboxProps } from './Checkbox.types'
