+++ src/components/checkboxGroup/CheckboxGroup.mdx
@@ -1,0 +1,17 @@
+import { ArgTypes, Canvas, Meta } from '@storybook/addon-docs'
+import * as CheckboxGroupStories from './CheckboxGroup.stories'
+import { CheckboxGroup } from '.'
+
+<Meta of={CheckboxGroupStories} component={CheckboxGroup} />
+
+# Checkbox
+
+사용자가 여러 개의 옵션 중 한 개 이상의 값을 선택할 때 사용한다.
+
+## 예시
+
+<Canvas of={CheckboxGroupStories.GroupOfCheckbox} />
+
+## API
+
+<ArgTypes of={CheckboxGroupStories} />
