+++ src/components/checkboxGroup/CheckboxGroup.types.ts
@@ -1,0 +1,18 @@
+import { CheckboxSize } from '../checkbox/Checkbox.types'
+
+export type OptionType = {
+  id: string
+  value: string
+  label: string
+  disabled?: boolean
+}
+
+export interface CheckboxGroupProps {
+  options: OptionType[]
+  name: string
+  label?: string
+  size?: CheckboxSize
+  direction?: 'horizontal' | 'vertical'
+  checkedOptions?: string[]
+  onChange?: (checkedOptions: string[]) => void
+}
