+++ src/components/checkboxGroup/index.tsx
@@ -1,0 +1,2 @@
+export { CheckboxGroup } from '../checkboxGroup/CheckboxGroup'
+export type { CheckboxGroupProps } from './CheckboxGroup.types'
