+++ src/index.tsx
@@ -5,3 +5,4 @@ export { Button } from './components/button'
 export { TextField } from './components/textfield'
 export { TextArea } from './components/textarea'
 export { Checkbox } from './components/checkbox'
+export { CheckboxGroup } from './components/checkboxGroup'
