+++ src/styles/index.scss
@@ -16,6 +16,7 @@
 @import './components/textfield';
 @import './components/textarea';
 @import './components/checkbox';
+@import './components/checkbox-group';
 
 // :root {
 //   --primary-color: 120, 60%;
